fix: position of the elements with negative viewport height
In the code provided for the `<ScrollControls />`, the position of foo is setted to a positive `viewport.height` value, which will be positioned upwards instead of downwards. Also, `viewport.height * 1` would leave the object at the same place as `viewport.height`, so it must be multiplied instead by the page number.